Move client routing into routing.cpp
